chore: revamp code and add temp messages while tests are being run
